requirements.txt
numpy
scipy
tqdm
lpips
# mmcv
imageio
imageio-ffmpeg
opencv-python
torch_efficient_distloss

